+++ src/Adult.cpp
@@ -1,8 +1,7 @@
 #include "Adult.h"
 
-Adult::Adult(string name,string address,long telephone_number,int age):Customer(name,address,telephone_number,age)
+Adult::Adult(string name,string address,long long telephone_number,int age):Customer(name,address,telephone_number,age)
 {
- //           Customer(name,address,telephone_number,age);
 
         save_int=SAVINGS_INTEREST;
         check_int=CHECK_INTEREST;
