+++ src/Student.cpp
@@ -1,7 +1,7 @@
 #include "Student.h"
 
 
-Student::Student(string name,string address,long telephone_number,int age):Customer(name,address,telephone_number,age){
+Student::Student(string name,string address,long long telephone_number,int age):Customer(name,address,telephone_number,age){
  //       Customer(name,address,telephone_number,age);
 
         save_int=SAVINGS_INTEREST;
